+++ lizard_portal/media/lizard_portal/lizard/model/AppScreen.js
@@ -1,5 +1,6 @@
 (function() {
-  " Under construction; just copied from Graph ";  Ext.define('Lizard.model.AppScreen', {
+  " Under construction; just copied from Graph ";
+  Ext.define('Lizard.model.AppScreen', {
     extend: 'Ext.data.Model',
     idProperty: 'slug',
     fields: [
@@ -30,4 +31,5 @@
       }
     }
   });
+
 }).call(this);