+++ lizard_portal/media/lizard_portal/lizard/store/AppScreen.js
@@ -1,4 +1,5 @@
 (function() {
+
   Ext.define('Lizard.store.AppScreen', {
     extend: 'Ext.data.Store',
     model: 'Lizard.model.AppScreen',
@@ -11,4 +12,5 @@
       }
     }
   });
+
 }).call(this);