+++ CHANGES.rst
@@ -2,8 +2,28 @@ Changelog of lizard-portal
 ==========================
 
 
-0.11.5 (unreleased)
--------------------
+0.15 (unreleased)
+-----------------
+
+- Nothing changed yet.
+
+
+0.14 (2012-03-08)
+-----------------
+
+- Implement initial support to view and validate configurations (#21).
+
+
+0.13 (2012-03-08)
+-----------------
+
+- Added first clickable layers in analysis screen. Still experimental.
+
+
+0.12 (2012-03-06)
+-----------------
+
+- Updated AppScreen.
 
 - Add addslashes filter to context in js template.